% Generated by build script
\name{fed_commercial_paper}
\alias{fed_commercial_paper}
\title{Commercial paper rates (daily)}
\usage{
fed_commercial_paper()
}
\arguments{
\item{lastobs}{Number of recent observations (default 250)}
\item{from}{Start date}
\item{to}{End date}
}
\value{tibble: date, series_code, value, description}
\description{Commercial paper rates (daily)}
