% Generated by build script
\name{fed_funds_rate}
\alias{fed_funds_rate}
\title{Federal funds rate (daily)}
\usage{
fed_funds_rate()
}
\arguments{
\item{lastobs}{Number of recent observations (default 250)}
\item{from}{Start date}
\item{to}{End date}
}
\value{tibble: date, series_code, value, description}
\description{Federal funds rate (daily)}
