% Generated by build script
\name{fed_h15}
\alias{fed_h15}
\title{All H.15 interest rates (Treasury + Fed funds + CP + prime)}
\usage{
fed_h15()
}
\arguments{
\item{lastobs}{Number of recent observations (default 60)}
}
\value{tibble: date, series_code, value, description}
\description{All H.15 interest rates (Treasury + Fed funds + CP + prime)}
