% Generated by build script
\name{fed_prime_rate}
\alias{fed_prime_rate}
\title{Bank prime loan rate}
\usage{
fed_prime_rate()
}
\arguments{
\item{lastobs}{Number of recent observations (default 250)}
\item{from}{Start date}
\item{to}{End date}
}
\value{tibble: date, series_code, value, description}
\description{Bank prime loan rate}
