% Generated by build script
\name{fed_treasury_yields}
\alias{fed_treasury_yields}
\title{Treasury yield curve rates (daily)}
\usage{
fed_treasury_yields()
}
\arguments{
\item{lastobs}{Number of most recent observations (default 250 = ~1yr)}
\item{from}{Start date (optional)}
\item{to}{End date (optional)}
}
\value{tibble: date, series_code, value, description}
\description{Treasury yield curve rates (daily)}
