% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fed.R
\name{fed_h15}
\alias{fed_h15}
\title{All H.15 interest rates (Treasury + Fed funds + CP + prime)}
\usage{
fed_h15(lastobs = 60)
}
\arguments{
\item{lastobs}{Number of recent observations (default 60)}
}
\value{
tibble: date, series_code, value, description
}
\description{
Convenience function that fetches all major H.15 series in one call.
Makes 4 HTTP requests.
}
